repo: phase 1 - move legacy facedancer code into deprecated
