+++ deprecated/SPIFlash.py
