+++ deprecated/legacy-applets/USBFtdi.py
