+++ deprecated/legacy-applets/USBKeyboard.py
