+++ deprecated/legacy-applets/USBMassStorage.py
