+++ deprecated/legacy-applets/USBQCEDL.py
