+++ deprecated/legacy-applets/USBSerial.py
