+++ deprecated/legacy-applets/USBSwitchTAS.py
