+++ deprecated/legacy-applets/facedancer-keyboard.py
