+++ facedancer/__init__.py
@@ -1,10 +1,9 @@
-
 import logging
 
 # Alias objects to make them easier to import.
 from .core import FacedancerUSBApp, FacedancerUSBHostApp, FacedancerBasicScheduler
 from .backends import *
-from .USBProxy import USBProxyFilter, USBProxyDevice
+from .deprecated.USBProxy import USBProxyFilter, USBProxyDevice
 
 from .devices import default_main as main
