+++ facedancer/backends/raspdancer.py
@@ -13,8 +13,8 @@ import logging
 
 from ..core import FacedancerApp
 from ..backends.MAXUSBApp import MAXUSBApp
-from ..USB import *
-from ..USBDevice import USBDeviceRequest
+from ..deprecated.USB import *
+from ..deprecated.USBDevice import USBDeviceRequest
 
 class RaspdancerMaxUSBApp(MAXUSBApp):
     app_name = "MAXUSB"
