+++ facedancer/deprecated/filters/standard.py
@@ -2,13 +2,15 @@
 # Standard filters for USBProxy that should (almost) always be used
 #
 
-import logging
+from ..USBProxy import USBProxyFilter
 
-from . import USBProxyFilter
-
-from ..             import *
-from ...descriptor  import *
-from ....errors     import *
+from ..USB import *
+from ..USBDevice import *
+from ..USBConfiguration import *
+from ..USBInterface import *
+from ..USBEndpoint import *
+from ..USBVendor import *
+from ..errors import *
 
 
 class USBProxySetupFilters(USBProxyFilter):
@@ -19,7 +21,7 @@ class USBProxySetupFilters(USBProxyFilter):
     RECIPIENT_DEVICE = 0
 
     DESCRIPTOR_DEVICE        = 0x01
-    DESCRIPTOR_CONFIGURATION = 0x02
+    DESCRIPTOR_CONFIRGUATION = 0x02
 
     MAX_PACKET_SIZE_EP0 = 64
 
@@ -45,19 +47,24 @@ class USBProxySetupFilters(USBProxyFilter):
             # If this is a configuration descriptor, store information relevant
             # to the configuration. We'll need this to set up the endpoint
             # hardware on the facedancer device.
-            if descriptor_type == self.DESCRIPTOR_CONFIGURATION and req.length >= 32:
+            if descriptor_type == self.DESCRIPTOR_CONFIRGUATION and req.length >= 32:
                 configuration = USBDescribable.from_binary_descriptor(data)
-                self.configurations[configuration.number] = configuration
-                logging.debug("-- Storing configuration {} --".format(configuration))
+                self.configurations[configuration.configuration_index] = configuration
+
+                if self.verbose > 1:
+                    print("-- Storing configuration {} --".format(configuration))
 
 
             if descriptor_type == self.DESCRIPTOR_DEVICE and req.length >= 7:
+
                 # Patch our data to overwrite the maximum packet size on EP0.
                 # See USBProxy.connect for a rationale on this.
                 device = USBDescribable.from_binary_descriptor(data)
                 device.max_packet_size_ep0 = 64
-                data = bytearray(device.get_descriptor())[:len(data)]
-                logging.debug("-- Patched device descriptor. --")
+                data = bytearray(device.get_descriptor(len(data)))
+
+                if self.verbose > 1:
+                    print("-- Patched device descriptor. --")
 
 
         return req, data, stalled
@@ -82,13 +89,14 @@ class USBProxySetupFilters(USBProxyFilter):
             if configuration_index in self.configurations:
                 configuration = self.configurations[configuration_index]
 
-                logging.debug("-- Applying configuration {} --".format(configuration))
+                if self.verbose > 0:
+                    print("-- Applying configuration {} --".format(configuration))
 
                 self.device.configured(configuration)
 
             # Otherwise, the host has applied a configuration without ever reading
             # its descriptor. This is mighty strange behavior!
-            else:
-                logging.warn("-- WARNING: Applying configuration {}, but we've never read that configuration's descriptor! --".format(configuration_index))
+            elif self.verbose > 0:
+                print("-- WARNING: Applying configuration {}, but we've never read that configuration's descriptor! --".format(configuration_index))
 
         return req, data
