+++ facedancer/filters/logging.py
@@ -3,7 +3,10 @@
 #
 
 import datetime
-from ..USBProxy import USBProxyFilter
+import logging
+
+from ..future.proxy       import USBProxyFilter
+
 
 class USBProxyPrettyPrintFilter(USBProxyFilter):
     """
@@ -25,14 +28,14 @@ class USBProxyPrettyPrintFilter(USBProxyFilter):
         """
 
         if self.verbose > 3 and req is None:
-            print("{} {}< --filtered out-- ".format(self.timestamp(), self.decoration))
+            logging.info("{} {}< --filtered out-- ".format(self.timestamp(), self.decoration))
             return req, data, stalled
 
         if self.verbose > 3:
-            print("{} {}{}".format(self.timestamp(), self.decoration, repr(req)))
+            logging.info("{} {}{}".format(self.timestamp(), self.decoration, repr(req)))
 
         if self.verbose > 3 and stalled:
-            print("{} {}< --STALLED-- ".format(self.timestamp(), self.decoration))
+            logging.info("{} {}< --STALLED-- ".format(self.timestamp(), self.decoration))
 
         if self.verbose > 4 and data:
             is_string = (req.request == 6) and (req.value >> 8 == 3)
@@ -49,11 +52,11 @@ class USBProxyPrettyPrintFilter(USBProxyFilter):
         # TODO: just call control_in, it's the same:
 
         if self.verbose > 3 and req is None:
-            print("{} {}> --filtered out-- ".format(self.timestamp(), self.decoration))
+            logging.info("{} {}> --filtered out-- ".format(self.timestamp(), self.decoration))
             return req, data
 
         if self.verbose > 3:
-            print("{} {}{}".format(self.timestamp(), self.decoration, repr(req)))
+            logging.info("{} {}{}".format(self.timestamp(), self.decoration, repr(req)))
 
         if self.verbose > 4 and data:
             self._pretty_print_data(data, '>', self.decoration)
@@ -67,9 +70,9 @@ class USBProxyPrettyPrintFilter(USBProxyFilter):
         """
         if self.verbose > 3 and req is None:
             if stalled:
-                print("{} {}> --STALLED-- ".format(self.timestamp(), self.decoration))
+                logging.info("{} {}> --STALLED-- ".format(self.timestamp(), self.decoration))
             else:
-                print("{} {}> --STALLED, but unstalled by filter-- ".format(self.timestamp(), self.decoration))
+                logging.info("{} {}> --STALLED, but unstalled by filter-- ".format(self.timestamp(), self.decoration))
 
         return req, data, stalled
 
@@ -109,7 +112,4 @@ class USBProxyPrettyPrintFilter(USBProxyFilter):
 
     def _pretty_print_data(self, data, direction_marker, decoration='', is_string=False, ep_marker=''):
         data = self._magic_decode(data) if is_string else bytes(data)
-        print("{} {}{}{}: {}".format(self.timestamp(), ep_marker, decoration, direction_marker, data))
-
-
-
+        logging.info("{} {}{}{}: {}".format(self.timestamp(), ep_marker, decoration, direction_marker, data))
